Update Pipfile to allow any Python version; add serialization rules to models in seed.py
The Pipfile has been updated to allow any version of Python to be used, instead of a specific version (3.8.13).

In seed.py, serialization rules have been added to the Game, Review, and User models. These rules specify which related objects should not be included when serializing the models. The Game model will not include reviews, the Review model will not include the associated game or user, and the User model will not include reviews.